Backend/requirements.txt
bert-extractive-summarizer==0.6.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
fastapi==0.63.0
filelock==3.0.12
fuzzywuzzy==0.18.0
h11==0.12.0
idna==2.10
joblib==1.0.0
murmurhash==1.0.5
nltk==3.5
numpy==1.19.5
packaging==20.8
plac==1.1.3
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
python-Levenshtein==0.12.0
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
spacy==2.3.5
srsly==1.0.5
starlette==0.13.6
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
tqdm==4.55.1
transformers==4.1.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
wasabi==0.8.0
